Don't implement IEnumerable to get foreach sugar.
Implementing IEnumerable signals a whole lot of other things that we
aren't really excited about, plus hangs all of the LINQ extensions and
so on off of this - but since we really only want to get a sort of "do
notation" for the Maybe struct via the foreach construct, we don't need
to implement IEnumerable: foreach is duck-typed, looking for a function
called GetEnumerator () that returns some type that has Current and
MoveNext () members defined.

Also adds a test case for the foreach syntax.